Refactor packages of UrlAlias and LocalizedUrlAliases, [mdz/webapps/internal/cudami/-/issues/169]
