+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/alias/LocalizedUrlAliasesMixIn.java
@@ -1,10 +1,10 @@
-package de.digitalcollections.model.jackson.mixin.alias;
+package de.digitalcollections.model.jackson.mixin.identifiable.alias;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.alias.LocalizedUrlAliases;
-import de.digitalcollections.model.alias.UrlAlias;
+import de.digitalcollections.model.identifiable.alias.LocalizedUrlAliases;
+import de.digitalcollections.model.identifiable.alias.UrlAlias;
 import java.util.List;
 import java.util.Locale;
