+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/alias/UrlAliasMixIn.java
@@ -1,8 +1,8 @@
-package de.digitalcollections.model.jackson.mixin.alias;
+package de.digitalcollections.model.jackson.mixin.identifiable.alias;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.alias.UrlAlias;
+import de.digitalcollections.model.identifiable.alias.UrlAlias;
 
 @JsonDeserialize(as = UrlAlias.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
